Hack: Copy every osc message on receive.
This fixes the memory corruption resulting from receiving more than one message at once. Before this, the messages were corrupted because the following messages wrote themselves to the same part of the memory before the earlier messages were processed. Now this is avoided, but not in a very elegant way.